LayoutTests/fast/images/text-recognition/image-overlay-block-with-newlines-expected.txt

PASS textarea.value is "Hello\nworld"
PASS successfullyParsed is true

TEST COMPLETE
